made changes to directory names from configuration - Config_Dir & files - Script_Dir
