+++ Config_Dir/myapp.conf
@@ -1,0 +1,10 @@
+[appdata] 
+enable = True
+ip = '192.168.195.182'
+username = 'tellabs'
+password = 'tellabs$123'
+file_name = 'shrey5local.conf'
+devUser = 'stack5shrey'
+devUserPwd = 'root'
+data_local = ['cmd0', 'cmd1', 'cmd2']
+data_local1 = "[[local|localrc]]\r\nRECLONE=True\r\nHOST_IP=192.168.195.182\r\nSERVICE_TOKEN=mytoken123\r\nADMIN_PASSWORD=openstack123\r\nMYSQL_PASSWORD=mysql123\r\nRABBIT_PASSWORD=rabbit123\r\nSERVICE_PASSWORD=$ADMIN_PASSWORD\r\nLOGFILE=$DEST/logs/stack.sh.log\r\nLOGDAYS=2\r\nenable_plugin neutron-lbaas https://github.com/openstack/neutron-lbaas.git stable/ocata\r\ndisable_service n-net c-api c-sch c-vol\r\nenable_service q-svc q-agt q-dhcp q-l3 q-meta q-lbaasv2"
