+++ Config_Dir/myappConfig.yaml
@@ -1,0 +1,38 @@
+# ================================================================
+# shrey2 Information
+# ================================================================
+shrey2:
+  - name: shrey
+    a1: A1
+    b1: B1
+    c1: C1
+    d1: D1
+    e1:
+      11: data11
+      12: data12
+  - name: shrey
+    a2: A2
+    b2: B2
+    c2: C2
+    d2: D2
+    e2:
+      21: data21
+      22: data22
+    # (Optional) Node Attribute JSON File
+    attribute_file: attribute
+
+# ================================================================
+# shrey1 Information
+# ================================================================
+shrey1:
+    name: shrey
+    a11: A11
+    b11: B11
+    c11: C11
+    d11: D11
+    e11:
+      111: data111
+      112: data112
+    f11:
+      211: data211
+      212: data212
