+++ Script_Dir/__init__.py
@@ -1,0 +1,8 @@
+try:
+    from sshAutomation import Devstack
+    from osloconfig import shrey
+    from readjson import get_jsonconfig
+    from readyaml import get_yamlconfig
+    from applogging import apploggingcompilationcheck
+except ImportError:
+    pass
