--- files/__init__.py
@@ -1,8 +1,0 @@
-try:
-    from sshAutomation import Devstack
-    from osloconfig import shrey
-    from readjson import get_jsonconfig
-    from readyaml import get_yamlconfig
-    from applogging import apploggingcompilationcheck
-except ImportError:
-    pass
